+++ frame/support/src/metadata.rs
@@ -469,11 +469,11 @@ mod tests {
 	}
 
 	impl event_module::Config for TestRuntime {
-		type Balance = u32;
+		type Balance = u128;
 	}
 
 	impl event_module2::Config for TestRuntime {
-		type Balance = u32;
+		type Balance = u128;
 	}
 
 	crate::parameter_types! {
